Update PLSS2LL.R
added as.numeric() to the condition. I have a column of section numbers coded as characters. This did not stop the function from running, but I got weird results (the resulting code was not in good format I guess). I'm not sure how this miss-specified operators work, but for example "14" > 9 returns FALSE.