Update PLSS2LL.R
A previous change added the first "as.numeric" in row 59, because when p$s was coded as a string, the conditional was returning erroneously. For x < 90, if x were coded as character (with or without a leading zero), x > 9 is FALSE. This meant that "04" got padded as well, resulting in a longer eventual plssid string. Testing it on a few coordinates, this extra padding can actually change the results from PLSS2LL. By adding the second "as.numeric" condition, the padding is done correctly regardless if p$s is character or numeric.